backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
aiosqlite==0.20.0
pyjwt==2.9.0
google-genai==1.5.0
python-multipart==0.0.12
pydantic==2.9.2
